feat: refactor job merging logic into job_utils and update monitoring section for improved clarity. end_time update when process finished
